Update README.md
Acted too quickly. Some feel that there might be some level of value with the content of this repo, so I'm reverting it back to where it was.

My personal thoughts are to abandon it, and eventually remove it as it has very low level of valuable content on it, it's quite repetitive, and basically shows my journey down the rabbit-hole that is Child Key Derivation.  I have learned so much in the last week as opposed to all the work and funds I put into it.

Frankly, the new repo is more apropos to the original project idea, has a simpler and more inclusive name, actually solves the problem in a simple and effective manner, and provides a higher level of value than this one. 

IOW, I'm ready to scrap this effort in favor of the new effort, which is actually solving the problem that this repo was supposed to do. Not to say that this repo is junk, just that it was a great staging area and learning experience, hopefully for all involved.

I welcome feedback to my opinions.

Bob