Update getBTC.conf
Changed first account name to "Default" to make it easy for getBTCInvoice to not have to choose an account.  See [getBTCInvoice] It doesn't really matter which one has the default name, and it's not required, but just makes it easier if someone only uses one account, or repeatedly uses a "default" account.