Update README.md
NOTE: Since you're using the shortcuts at the bottom, it's not necessary to use the [getBTCAddress][getBTCAddress] format and instead can simply use [getBTCAddress] for those links.  See the link for [getBTC.conf] in the table of contents.  Great idea for doing that.  I think we have to do the same on every page so that there is consistency.  Easy enough to do.  REALLY nice looking now.