Create setBTCPrice.php
This is the alternative to using the getBTCRate function in getBTC.php. It stores the current price from one of three accessible exchanges in a JSON file for use by PHP on the server, which eliminates potential rate limits from the servers.
It is triggered to run by a system cron job on the server, once each minute. See line 5 of this file for the cron job to add to roots crontab.